Tinkering and Regret
I regret thinking that working with the Device Trees was a good idea. Obviously I need some way of determing what board the kernel is running on on bootup and this is generally the way to go, but its killing me how to do this. Nothing is clear! ATAGS may just be the way to go, but the device tree system seems to be the modern way.