Trees and Restructuring
Added a barebones handwritin device tree for the RPI 3b+ (YAY). Im just so glad thats done with and I am starting to understand exactly what I am describing. I mapped in all of the peripherals into childnodes of root.

Because of this, I came up with a MUCH better way of organizing my drivers and came up with a simple naming scheme. No more complex "oh but you have to implement *THIS* api!!!" bull. Through the use of the data tree I can specify which drivers to use when so its a fat yeet. The future is bright.