Driver Fixes  + Updating Architecture
Hopefully moving the tools into the project but that kinda thing is proving tough.

I updated the driver architecture a bit to be more flexible. The old way of doing things was a tiny bit silly and needed this refinement. I still need to go through and set the subroutines to use the base address specified in the driver header.

Speaking of which, I came up with a spec for how headers for drivers should be formatted. I believe this will be the last iteration on the general structure. The first address holds an identifier. This can be set to 0x0 as an invalid state. The second address holds the last address of the driver giving the size of said driver. The third address holds the base address for the driver. After that is the subroutine lookup table - but that's not new.

I need to better specify how to notate the subroutines but the basic idea is there.